# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalDecision": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ApprovalAction"
          },
          "actor_id": {
            "format": "uuid",
            "title": "Actor Id",
            "type": "string"
          },
          "actor_role": {
            "$ref": "#/components/schemas/Role"
          },
          "approved_draft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved Draft Id"
          },
          "approved_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved Note"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "diff": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diff"
          },
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "original_note": {
            "title": "Original Note",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "$ref": "#/components/schemas/ApprovalStatus"
          },
          "workflow_id": {
            "format": "uuid",
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workflow_id",
          "draft_id",
          "actor_id",
          "actor_role",
          "action",
          "status",
          "original_note",
          "approved_note",
          "approved_draft_id",
          "reason",
          "diff",
          "created_at"
        ],
        "title": "ApprovalDecision",
        "type": "object"
      },
      "ApprovalStatus": {
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED"
        ],
        "title": "ApprovalStatus",
        "type": "string"
      },
      "ApproveRequest": {
        "additionalProperties": false,
        "properties": {
          "draft_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Draft Id",
            "type": "string"
          }
        },
        "required": [
          "draft_id"
        ],
        "title": "ApproveRequest",
        "type": "object"
      },
...
      "DecisionResponse": {
        "properties": {
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          },
          "review": {
            "$ref": "#/components/schemas/ReviewResponse"
          }
        },
        "required": [
          "review",
          "replayed"
        ],
        "title": "DecisionResponse",
        "type": "object"
      },
...
      "EditAndApproveRequest": {
        "additionalProperties": false,
        "properties": {
          "draft_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Draft Id",
            "type": "string"
          },
          "edited_note": {
            "maxLength": 64000,
            "minLength": 1,
            "title": "Edited Note",
            "type": "string"
          }
        },
        "required": [
          "draft_id",
          "edited_note"
        ],
        "title": "EditAndApproveRequest",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "pattern": "^[A-Z_]+$",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "description": "Safe human-readable explanation; never raw server details.",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JobState": {
        "enum": [
          "PENDING",
          "QUEUED",
          "STARTED",
          "COMPLETED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "JobState",
        "type": "string"
      },
...
      "RejectRequest": {
        "additionalProperties": false,
        "properties": {
          "draft_id": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Draft Id",
            "type": "string"
          },
          "reason": {
            "maxLength": 4000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "draft_id",
          "reason"
        ],
        "title": "RejectRequest",
        "type": "object"
      },
...
      "ReviewResponse": {
        "properties": {
          "approval_allowed": {
            "title": "Approval Allowed",
            "type": "boolean"
          },
          "approval_status": {
            "$ref": "#/components/schemas/ApprovalStatus"
          },
          "cancellation_requested": {
            "title": "Cancellation Requested",
            "type": "boolean"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "draft": {
            "$ref": "#/components/schemas/ClinicalNoteDraft"
          },
          "finalization_request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinalizationRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "job_state": {
            "$ref": "#/components/schemas/JobState"
          },
          "safety_verdict": {
            "$ref": "#/components/schemas/SafetyVerdict"
          },
          "workflow_id": {
            "format": "uuid",
            "title": "Workflow Id",
            "type": "string"
          },
          "workflow_state": {
            "title": "Workflow State",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "job_id",
          "workflow_state",
          "job_state",
          "cancellation_requested",
          "approval_status",
          "draft",
          "safety_verdict",
          "approval_allowed",
          "decision",
          "finalization_request"
        ],
        "title": "ReviewResponse",
        "type": "object"
      },
...
      "StartWorkflowRequest": {
        "additionalProperties": false,
        "properties": {
          "case_summary": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Case Summary",
            "type": "string"
          },
          "clinical_question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Clinical Question",
            "type": "string"
          },
          "patient_context": {
            "default": "",
            "maxLength": 4000,
            "title": "Patient Context",
            "type": "string"
          }
        },
        "required": [
          "clinical_question",
          "case_summary"
        ],
        "title": "StartWorkflowRequest",
        "type": "object"
      },
...
      "WorkflowAcceptedResponse": {
        "properties": {
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          },
          "workflow_id": {
            "format": "uuid",
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "state",
          "status_url",
          "workflow_id"
        ],
        "title": "WorkflowAcceptedResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "JWT access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Domain Copilot HTTP API. Bearer JWT identity and stored roles are checked server-side. See docs/API-CONTRACTS.md for ownership and SSE semantics.",
    "title": "Domain Copilot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runs": {
      "post": {
        "description": "Start the registered clinical workflow. The worker creates its durable run.\n\nRoles: analyst, reviewer, admin. Owner only for analyst; reviewer/admin may view all runs. Identity and current role come from the validated JWT and stored user.",
        "operationId": "start_workflow_api_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "DOMAIN_ERROR, APPLICATION_ERROR or HTTP_ERROR: invalid operation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_AUTHENTICATED. WWW-Authenticate: Bearer.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PERMISSION_DENIED or RESOURCE_FORBIDDEN."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RESOURCE_NOT_FOUND or HTTP_ERROR."
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HTTP_ERROR: method not allowed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_STATE_TRANSITION: conflict with persisted state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UPLOAD_TOO_LARGE."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR or INVARIANT_VIOLATION: invalid input."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR or CONFIGURATION_ERROR. Static, sanitized detail."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JOB_STORE_UNAVAILABLE, KNOWLEDGE_UNAVAILABLE, APPROVAL_STORE_UNAVAILABLE or HISTORY_STORE_UNAVAILABLE. Static, sanitized detail."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Start Workflow",
        "tags": [
          "workflows"
        ],
        "x-ownership": "Owner only for analyst; reviewer/admin may view all runs.",
        "x-roles": [
          "analyst",
          "reviewer",
          "admin"
        ]
      }
    },
...
    "/api/v1/runs/{workflow_id}/approval": {
      "get": {
        "description": "Roles: reviewer, admin. VIEW_ALL_RUNS permits cross-owner clinical review. Identity and current role come from the validated JWT and stored user.",
        "operationId": "review_draft_api_v1_runs__workflow_id__approval_get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "DOMAIN_ERROR, APPLICATION_ERROR or HTTP_ERROR: invalid operation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_AUTHENTICATED. WWW-Authenticate: Bearer.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PERMISSION_DENIED or RESOURCE_FORBIDDEN."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RESOURCE_NOT_FOUND or HTTP_ERROR."
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HTTP_ERROR: method not allowed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_STATE_TRANSITION: conflict with persisted state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UPLOAD_TOO_LARGE."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR or INVARIANT_VIOLATION: invalid input."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR or CONFIGURATION_ERROR. Static, sanitized detail."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JOB_STORE_UNAVAILABLE, KNOWLEDGE_UNAVAILABLE, APPROVAL_STORE_UNAVAILABLE or HISTORY_STORE_UNAVAILABLE. Static, sanitized detail."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Review Draft",
        "tags": [
          "approvals"
        ],
        "x-ownership": "VIEW_ALL_RUNS permits cross-owner clinical review.",
        "x-roles": [
          "reviewer",
          "admin"
        ]
      }
    },
    "/api/v1/runs/{workflow_id}/approval/approve": {
      "post": {
        "description": "Roles: reviewer, admin. VIEW_ALL_RUNS permits cross-owner clinical review. Identity and current role come from the validated JWT and stored user.",
        "operationId": "approve_draft_api_v1_runs__workflow_id__approval_approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "DOMAIN_ERROR, APPLICATION_ERROR or HTTP_ERROR: invalid operation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_AUTHENTICATED. WWW-Authenticate: Bearer.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PERMISSION_DENIED or RESOURCE_FORBIDDEN."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RESOURCE_NOT_FOUND or HTTP_ERROR."
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HTTP_ERROR: method not allowed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_STATE_TRANSITION: conflict with persisted state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UPLOAD_TOO_LARGE."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR or INVARIANT_VIOLATION: invalid input."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR or CONFIGURATION_ERROR. Static, sanitized detail."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JOB_STORE_UNAVAILABLE, KNOWLEDGE_UNAVAILABLE, APPROVAL_STORE_UNAVAILABLE or HISTORY_STORE_UNAVAILABLE. Static, sanitized detail."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Approve Draft",
        "tags": [
          "approvals"
        ],
        "x-ownership": "VIEW_ALL_RUNS permits cross-owner clinical review.",
        "x-roles": [
          "reviewer",
          "admin"
        ]
      }
    },
    "/api/v1/runs/{workflow_id}/approval/edit-and-approve": {
      "post": {
        "description": "Roles: reviewer, admin. VIEW_ALL_RUNS permits cross-owner clinical review. Identity and current role come from the validated JWT and stored user.",
        "operationId": "edit_and_approve_draft_api_v1_runs__workflow_id__approval_edit_and_approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditAndApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "DOMAIN_ERROR, APPLICATION_ERROR or HTTP_ERROR: invalid operation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_AUTHENTICATED. WWW-Authenticate: Bearer.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PERMISSION_DENIED or RESOURCE_FORBIDDEN."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RESOURCE_NOT_FOUND or HTTP_ERROR."
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HTTP_ERROR: method not allowed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_STATE_TRANSITION: conflict with persisted state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UPLOAD_TOO_LARGE."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR or INVARIANT_VIOLATION: invalid input."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR or CONFIGURATION_ERROR. Static, sanitized detail."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JOB_STORE_UNAVAILABLE, KNOWLEDGE_UNAVAILABLE, APPROVAL_STORE_UNAVAILABLE or HISTORY_STORE_UNAVAILABLE. Static, sanitized detail."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Edit And Approve Draft",
        "tags": [
          "approvals"
        ],
        "x-ownership": "VIEW_ALL_RUNS permits cross-owner clinical review.",
        "x-roles": [
          "reviewer",
          "admin"
        ]
      }
    },
    "/api/v1/runs/{workflow_id}/approval/reject": {
      "post": {
        "description": "Roles: reviewer, admin. VIEW_ALL_RUNS permits cross-owner clinical review. Identity and current role come from the validated JWT and stored user.",
        "operationId": "reject_draft_api_v1_runs__workflow_id__approval_reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "DOMAIN_ERROR, APPLICATION_ERROR or HTTP_ERROR: invalid operation."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "NOT_AUTHENTICATED. WWW-Authenticate: Bearer.",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "const": "Bearer",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PERMISSION_DENIED or RESOURCE_FORBIDDEN."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "RESOURCE_NOT_FOUND or HTTP_ERROR."
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "HTTP_ERROR: method not allowed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_STATE_TRANSITION: conflict with persisted state."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "UPLOAD_TOO_LARGE."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR or INVARIANT_VIOLATION: invalid input."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR or CONFIGURATION_ERROR. Static, sanitized detail."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JOB_STORE_UNAVAILABLE, KNOWLEDGE_UNAVAILABLE, APPROVAL_STORE_UNAVAILABLE or HISTORY_STORE_UNAVAILABLE. Static, sanitized detail."
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Reject Draft",
        "tags": [
          "approvals"
        ],
        "x-ownership": "VIEW_ALL_RUNS permits cross-owner clinical review.",
        "x-roles": [
          "reviewer",
          "admin"
        ]
      }
    },
...
  }
}